% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_core.R
\name{dwt_periodic}
\alias{dwt_periodic}
\alias{idwt_periodic}
\title{Periodic discrete wavelet transform}
\usage{
dwt_periodic(x, L, filter = "coif1")

idwt_periodic(w, L, filter = "coif1")
}
\arguments{
\item{x}{numeric vector.}

\item{L}{number of levels.}

\item{filter}{see \code{\link[=wavelet_filter]{wavelet_filter()}}.}

\item{w}{wavelet coefficients from \code{\link[=dwt_periodic]{dwt_periodic()}}.}
}
\value{
numeric vector of the same length as \code{x}.
}
\description{
Forward (\code{dwt_periodic()}) and inverse (\code{idwt_periodic()}) periodised
orthonormal DWT with \code{L} levels, as in Vidakovic (1999). \code{length(x)} must be
divisible by \code{2^L}. Coefficients are returned as
\code{c(smooth, detail_coarsest, ..., detail_finest)}. Translated from the
MATLAB routines \code{dwtr} and \code{Idwtr} of B. Vidakovic.
}
\examples{
x <- rnorm(32)
all.equal(idwt_periodic(dwt_periodic(x, 3, "coif1"), 3, "coif1"), x)
}
