% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpa_density_ci.R
\name{plot.fpa_density_ci}
\alias{plot.fpa_density_ci}
\title{Plot a GPV density estimate with confidence bands}
\usage{
\method{plot}{fpa_density_ci}(x, true = NULL, ...)
}
\arguments{
\item{x}{an \code{"fpa_density_ci"} object.}

\item{true}{optional true density function for simulated data.}

\item{...}{passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot a GPV density estimate with confidence bands
}
