% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpa_density.R
\name{plot.fpa_density}
\alias{plot.fpa_density}
\title{Plot an estimated private-value density}
\usage{
\method{plot}{fpa_density}(x, true = NULL, rug = TRUE, ...)
}
\arguments{
\item{x}{an \code{"fpa_density"} object.}

\item{true}{optional true density function for simulated data.}

\item{rug}{logical; add a rug of the pseudo-values.}

\item{...}{passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\description{
Plot an estimated private-value density
}
