% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpa_values.R
\name{plot.fpa_values}
\alias{plot.fpa_values}
\title{Plot recovered private values against bids}
\usage{
\method{plot}{fpa_values}(x, true = NULL, ...)
}
\arguments{
\item{x}{an \code{"fpa_values"} object.}

\item{true}{optional true values (vector or matrix in the layout of \code{bids})
for simulated data.}

\item{...}{passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\description{
Plot recovered private values against bids
}
