% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdf.R
\name{plot.qdf}
\alias{plot.qdf}
\title{Plot a quantile density estimate}
\usage{
\method{plot}{qdf}(x, true = NULL, ...)
}
\arguments{
\item{x}{a \code{"qdf"} object.}

\item{true}{optional function of \code{u} giving the true quantile density, drawn
for comparison.}

\item{...}{further arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot a quantile density estimate
}
